// generated/src/app/api-models/claims-duplicate-rule.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimsDuplicateRule {
  claimDupRule: string;
  claimType: string;
  description: string;
  primDtUsed1: string;
  primDtDaysBef1: number;
  primDtDaysAft1: number;
  sameSrvProv1: string;
  sameTotBill1: string;
  samePlcServ1: string;
  specDateUsed1: string;
  specDtDaysBef1: number;
  specDtDaysAft1: number;
  sameBillAmt1: string;
  sameProcCode1: string;
  sameMedDef1: string;
  action1: string;
  reasonCode1: string;
  primDtUsed2: string;
  primDtDaysBef2: number;
  primDtDaysAft2: number;
  sameSrvProv2: string;
  sameTotBill2: string;
  samePlcServ2: string;
  specDateUsed2: string;
  specDtDaysBef2: number;
  specDtDaysAft2: number;
  sameBillAmt2: string;
  sameProcCode2: string;
  sameMedDef2: string;
  action2: string;
  reasonCode2: string;
  primDtUsed3: string;
  primDtDaysBef3: number;
  primDtDaysAft3: number;
  sameSrvProv3: string;
  sameTotBill3: string;
  samePlcServ3: string;
  specDateUsed3: string;
  specDtDaysBef3: number;
  specDtDaysAft3: number;
  sameBillAmt3: string;
  sameProcCode3: string;
  sameMedDef3: string;
  action3: string;
  reasonCode3: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  sameProcMod1: string;
  sameProcMod2: string;
  sameProcMod3: string;
  sameTooth1: string;
  sameTooth2: string;
  sameTooth3: string;
}